Simplifying refactor of LinedDisplay
The LinedDisplay abstraction was requiring dynamic reallocation of lines
based on line configuration and was requiring more moving parts, at which
point I realized that, much like the LEDs, much of it was not necessary.
This change removes some of those moving parts and tailors LinedDIsplay
to this specific project.